Exclude cypress from tsconfig
